Fix INT32 overflow in threshold calculations (phase3 and common)
- Changed CAST(... AS INTEGER) to CAST(... AS BIGINT) for all threshold calculations
- Threshold calculation multiplies needed * 10000 which can exceed INT32
- Even though final threshold is 0-10000, intermediate calculation can overflow
- Fixed 6 locations:
  * phase3_cohort_creation.py: 4 locations (2 in sample_threshold CTE, 2 in WHERE clause)
  * common.py: 2 locations (sample_threshold CTE in ensure_cohort_views)
- Remaining CAST(event_year AS INTEGER) are safe (small integers: 2016-2019)
- Fixes: Conversion Error: Type DOUBLE with value 30149350000.0 can't be cast to INT32 at LINE 133

+++ 2_create_cohort/phases/common.py
@@ -678,7 +678,7 @@ def ensure_cohort_views(conn, logger):
             ),
             sample_threshold AS (
                 SELECT 
-                    CAST(ROUND((SELECT needed FROM needed_count)::DOUBLE / GREATEST((SELECT available FROM available_controls), 1) * 10000) AS INTEGER) as threshold
+                    CAST(ROUND((SELECT needed FROM needed_count)::DOUBLE / GREATEST((SELECT available FROM available_controls), 1) * 10000) AS BIGINT) as threshold
             )
             SELECT 
                 mi_person_key
@@ -762,7 +762,7 @@ def ensure_cohort_views(conn, logger):
             ),
             sample_threshold AS (
                 SELECT 
-                    CAST(ROUND((SELECT needed FROM needed_count)::DOUBLE / GREATEST((SELECT available FROM available_controls), 1) * 10000) AS INTEGER) as threshold
+                    CAST(ROUND((SELECT needed FROM needed_count)::DOUBLE / GREATEST((SELECT available FROM available_controls), 1) * 10000) AS BIGINT) as threshold
             )
             SELECT 
                 mi_person_key
